@@ -1,20 +1,15 @@
 package net.proys.proysrail.Adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.CompoundButton;
 import android.widget.RadioButton;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import net.proys.proysrail.Items.FuelMainItem;
 import net.proys.proysrail.Items.PuantajDetayItem;
 import net.proys.proysrail.R;
-
-import org.w3c.dom.Text;
 
 import java.util.ArrayList;
 
