@@ -1,12 +1,11 @@
 package net.proys.proysrail.Adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import net.proys.proysrail.Items.PuantajMainItem;
 import net.proys.proysrail.R;
