@@ -2,14 +2,13 @@ package net.proys.proysrail.Adapters;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import net.proys.proysrail.Fragments.MakinePuantaj;
 import net.proys.proysrail.Items.MakinePopupItem;
 import net.proys.proysrail.Makine_Isletme_MainFrame;
 import net.proys.proysrail.R;
