@@ -1,18 +1,13 @@
 package net.proys.proysrail.Fragments;
 
-import android.content.Context;
-import android.content.Intent;
-import android.graphics.Color;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import net.proys.proysrail.Adapters.PuantajMainAdapter;
 import net.proys.proysrail.Adapters.makine_isletme_popup_adapter;
 import net.proys.proysrail.Items.MakinePopupItem;
 import net.proys.proysrail.R;
