@@ -1,33 +1,19 @@
 package net.proys.proysrail;
 
-import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Intent;
-import android.graphics.Rect;
-import android.os.Handler;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.fragment.app.FragmentTransaction;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
-import android.util.Log;
-import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
-import android.view.Window;
-import android.view.WindowManager;
-import android.widget.Adapter;
-import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.ListView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import net.proys.proysrail.Adapters.PuantajMainAdapter;
 import net.proys.proysrail.Adapters.makine_isletme_popup_adapter;
 import net.proys.proysrail.Fragments.MakineFault;
 import net.proys.proysrail.Fragments.MakineFuel;
@@ -35,12 +21,6 @@ import net.proys.proysrail.Fragments.MakinePopup;
 import net.proys.proysrail.Fragments.MakinePuantaj;
 import net.proys.proysrail.Fragments.MakineStatement;
 import net.proys.proysrail.Internfaces.Visibility;
-import net.proys.proysrail.Items.MakinePopupItem;
-
-import java.util.ArrayList;
-
-import static net.proys.proysrail.R.drawable.background;
-import static net.proys.proysrail.R.drawable.navbar_back;
 
 public class Makine_Isletme_MainFrame extends AppCompatActivity  implements Visibility {
    RelativeLayout ustbar;
